# app/src/api/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "JInshanBodyShopApi",
    "version": "1.0"
  },
  "paths": {
...
    "/Ovh": {
      "post": {
        "tags": ["Ovh"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PresignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/download": {
      "get": {
        "tags": ["Ovh"],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/GetProductsByIds": {
      "get": {
        "tags": ["Product"],
        "parameters": [
          {
            "name": "productIds[]",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/GetProducts": {
      "get": {
        "tags": ["Product"],
        "parameters": [
          {
            "name": "filterByCategories[]",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ProductType"
              }
            }
          },
          {
            "name": "filterByWithdrawn",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "shop",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProductShop"
            }
          },
          {
            "name": "filterByName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderByColumn",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "productid"
            }
          },
          {
            "name": "orderAscending",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductApiGetReturnType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductApiGetReturnType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductApiGetReturnType"
                }
              }
            }
          }
        }
      }
    },
    "/Product": {
      "get": {
        "tags": ["Product"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Product"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostProduct"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostProduct"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostProduct"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Product"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostProduct"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostProduct"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostProduct"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": ["Product"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/GetSimilarProducts": {
      "get": {
        "tags": ["SimilarProducts"],
        "parameters": [
          {
            "name": "productId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/SimilarProducts": {
      "post": {
        "tags": ["SimilarProducts"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostSimilarProducts"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostSimilarProducts"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostSimilarProducts"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/DeleteSimilarProducts": {
      "delete": {
        "tags": ["SimilarProducts"],
        "parameters": [
          {
            "name": "productId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "similarProductId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PostProduct": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "withdrawn": {
            "type": "boolean",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/ProductType"
          },
          "shop": {
            "$ref": "#/components/schemas/ProductShop"
          }
        },
        "additionalProperties": false
      },
      "PostSimilarProducts": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "similarProductId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "PresignRequest": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Product": {
        "required": ["category"],
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "number",
            "format": "double"
          },
          "category": {
            "$ref": "#/components/schemas/ProductType"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "withdrawn": {
            "type": "boolean"
          },
          "shop": {
            "$ref": "#/components/schemas/ProductShop"
          }
        },
        "additionalProperties": false
      },
...
      "ProductApiGetReturnType": {
        "required": ["items"],
        "type": "object",
        "properties": {
          "maxPages": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductShop": {
        "enum": [0, 1, 2],
        "type": "integer",
        "format": "int32"
      },
      "ProductType": {
        "enum": [0, 1, 2, 3, 4, 5, 6, 7],
        "type": "integer",
        "format": "int32"
      },
...
    }
  }
}